Se termina implementacion de funcionalidades pedidas en el frontend.

+++ Lab1_TBD/src/main/java/Lab1_TBD/controllers/EmergencyController.java
@@ -36,7 +36,7 @@ public class EmergencyController {
     }
 
     @GetMapping("/emergencies/volunteersInRadius/{id}/{radius}")
-    public ResponseEntity<List<VolunteerEntity>> getVolunteersInEmergencyRadius(@PathVariable Long id, @PathVariable Long radius) {
+    public ResponseEntity<List<VolunteerEntity>> getVolunteersInEmergencyRadius(@PathVariable Long id, @PathVariable Double radius) {
         List<VolunteerEntity> volunteers = emergencyService.getVolunteersInEmergencyRadius(id, radius);
         return ResponseEntity.ok(volunteers);
     }
