+++ Lab1_TBD/src/main/java/Lab1_TBD/services/EmergencyService.java
@@ -18,7 +18,7 @@ public class EmergencyService {
     public EmergencyEntity saveEmergency(EmergencyEntity emergency) { return emergencyRepository.save(emergency); }
     public EmergencyEntity getEmergencyById(Long id) { return emergencyRepository.findById(id); }
 
-    public List<VolunteerEntity> getVolunteersInEmergencyRadius(Long id, Long radius) { return emergencyRepository.getVolunteersInEmergencyRadius(id, radius); }
+    public List<VolunteerEntity> getVolunteersInEmergencyRadius(Long id, Double radius) { return emergencyRepository.getVolunteersInEmergencyRadius(id, radius); }
     public EmergencyEntity updateEmergency(EmergencyEntity emergency) { return emergencyRepository.update(emergency); }
     public boolean deleteEmergency(Long id) { return emergencyRepository.deleteById(id); }
     public List<EmergencyEntity> getAll(String token) {
